+++ src/app/common/services/storage/storage.ts
@@ -4,6 +4,7 @@ export const enum StorageKey {
   XpTrackerRecentPlayers = 'xp-tracker-recent-players',
   XpTrackerFavoritePlayers = 'xp-tracker-favorite-players',
   XpTrackerViewType = 'xp-tracker-view-type',
+  XpTrackerScrapingOffset = 'xp-tracker-scraping-offset',
 
   PriceTrackerRecentItems = 'price-tracker-recent-items',
   PriceTrackerFavoriteItems = 'price-tracker-favorite-items',